requirements.txt
# Audio capture & playback
sounddevice==0.5.1
numpy==1.26.4

# Voice Activity Detection — uses pure-Python energy-based detection (no C compiler needed)

# OpenAI SDK (Whisper STT, GPT Chat, TTS)
openai==1.61.0

# Environment variable loading
python-dotenv==1.0.1

# MP3 decoding for TTS playback
pydub==0.25.1

# pydub requires ffmpeg — install via:
#   Windows: choco install ffmpeg  OR  download from https://ffmpeg.org

